Test the vsix, not the build layout (#2156)
The VSCode C# Extension build process follows the VS Code docs and runs tests directly inside of its repo root. This unfortunately gives a false sense of security because bugs can be introduced during VSIX packaging [particularly due to missing content in node_modules or excluded via .vscodeignore].

This change addresses this problem by moving our CI tests to execute the VSIX instead of the build's intermediate artifacts. Specifically:

build the vsix
unpackage the vsix
instrument the unpackaged vsix
run tests with VS Code Host pointing to the unpackaged vsix
This makes our CI tests ~= to the user's runtime experience and will greatly help us with size reduction efforts.

To support this change, I also moved our build system from package.json to Gulp. This makes the build scripts significantly easier to understand, provides intellisense for build scripts, and build-time type checking for their contents.

I also strengthened the repo's use of .vscodeignore by creating a copy of the file for each scenario [online packages and offline packages]. The new gulp packaging scripts take advantage of these files to produce packages with predictable contents regardless of when packaging occurs. [small caveat, @akshita31 will be adding a test that validates that net-new content does not start sneaking into the vsix package].

+++ test/coverageWritingTestRunner.ts
@@ -6,7 +6,6 @@
 'use strict';
 
 import shelljs = require("async-shelljs");
-
 import path = require('path');
 const fs = require('async-file');
 import Mocha = require('mocha');
@@ -46,11 +45,7 @@ export default class CoverageWritingTestRunner {
 
     private async writeCoverage(): Promise<void> {
         if (typeof __coverage__ !== 'undefined') {
-            let nycFolderPath = path.join(process.env.CODE_EXTENSIONS_PATH, ".nyc_output", "integration");
-
-            if (!(await fs.exists(nycFolderPath))) {
-                await fs.mkdir(nycFolderPath);
-            }
+            let nycFolderPath = path.join(process.env.CODE_WORKSPACE_ROOT, ".nyc_output", "integration");
 
             let rawCoverageJsonPath: string;
             let remappedCoverageJsonPath: string;
@@ -59,51 +54,56 @@ export default class CoverageWritingTestRunner {
             let nodePath: string;
 
             try {
-                rawCoverageJsonPath = path.join(nycFolderPath, `${process.env.OSVC_SUITE}.json.raw`);
-                remappedCoverageJsonPath = path.join(nycFolderPath, `${process.env.OSVC_SUITE}.json`);
-                outFolderPath = path.join(process.env.CODE_EXTENSIONS_PATH, "out");
-                remapIstanbulPath = path.join(process.env.CODE_EXTENSIONS_PATH, "node_modules", "remap-istanbul", "bin", "remap-istanbul.js");
-                nodePath = "";
-                if (process.env.NVM_BIN) {
-                    nodePath = `${process.env.NVM_BIN}${path.sep}`;
-                }
+                if (!(await fs.exists(nycFolderPath))) {
+                    await fs.mkdirp(nycFolderPath);
 
-                await fs.writeTextFile(rawCoverageJsonPath, JSON.stringify(__coverage__));
-                
-                let result = await shelljs.asyncExec(`${nodePath}node ${remapIstanbulPath} -i ${rawCoverageJsonPath} -o ${remappedCoverageJsonPath}`, {
-                    cwd: outFolderPath
-                });
-
-                
-
-                let remappedResult = JSON.parse(await fs.readTextFile(remappedCoverageJsonPath));
-
-                let finalResult = {};
-
-                for (let key in remappedResult) {
-                    if (remappedResult[key].path) {
-                        let realPath = key.replace("../", "./");
-
-                        finalResult[realPath] = remappedResult[key];
-                        finalResult[realPath].path = realPath;
+                    rawCoverageJsonPath = path.join(nycFolderPath, `${process.env.OSVC_SUITE}.json.raw`);
+                    remappedCoverageJsonPath = path.join(nycFolderPath, `${process.env.OSVC_SUITE}.json`);
+                    outFolderPath = path.join(process.env.CODE_WORKSPACE_ROOT, "out");
+                    remapIstanbulPath = path.join(process.env.CODE_WORKSPACE_ROOT, "node_modules", "remap-istanbul", "bin", "remap-istanbul.js");
+                    nodePath = "";
+                    if (process.env.NVM_BIN) {
+                        nodePath = `${process.env.NVM_BIN}${path.sep}`;
                     }
-                    else {
-                        finalResult[key] = remappedResult[key];
+
+                    await fs.writeTextFile(rawCoverageJsonPath, JSON.stringify(__coverage__));
+
+                    let result = await shelljs.asyncExec(`${nodePath}node ${remapIstanbulPath} -i ${rawCoverageJsonPath} -o ${remappedCoverageJsonPath}`, {
+                        cwd: outFolderPath
+                    });
+
+
+
+                    let remappedResult = JSON.parse(await fs.readTextFile(remappedCoverageJsonPath));
+
+                    let finalResult = {};
+
+                    for (let key in remappedResult) {
+                        if (remappedResult[key].path) {
+                            let realPath = key.replace("../", "./");
+
+                            finalResult[realPath] = remappedResult[key];
+                            finalResult[realPath].path = realPath;
+                        }
+                        else {
+                            finalResult[key] = remappedResult[key];
+                        }
                     }
+
+                    await fs.writeTextFile(remappedCoverageJsonPath, JSON.stringify(finalResult));
+
+                    console.log(`done remapping ${finalResult}`);
                 }
-
-                await fs.writeTextFile(remappedCoverageJsonPath, JSON.stringify(finalResult));
-
-                console.log(`done remapping ${finalResult}`);
             }
             catch (e) {
-                console.log(`Coverage remapping failure: ${JSON.stringify(e)}`);
-                console.log(`* rawCoverageJsonPath: ${rawCoverageJsonPath}`);
-                console.log(`* remappedCoverageJsonPath: ${remappedCoverageJsonPath}`);
-                console.log(`* outFolderPath: ${outFolderPath}`);
-                console.log(`* remapIstanbulPath: ${remapIstanbulPath}`);
-                console.log(`* nodePath: ${nodePath}`);
+                    console.log(`Coverage remapping failure: ${JSON.stringify(e)}`);
+                    console.log(`* rawCoverageJsonPath: ${rawCoverageJsonPath}`);
+                    console.log(`* remappedCoverageJsonPath: ${remappedCoverageJsonPath}`);
+                    console.log(`* outFolderPath: ${outFolderPath}`);
+                    console.log(`* remapIstanbulPath: ${remapIstanbulPath}`);
+                    console.log(`* nodePath: ${nodePath}`);
+                    console.log(e);
+                }
             }
-        }
     }
-}
+    }
